Add edit contact modal
