+++ src/components/Contacts/EditModal.module.css
@@ -7,7 +7,7 @@
     z-index: 1;
     background: rgba(0, 0, 0, 0.25);
   }
-  
+
   .modal {
     position: fixed;
     top: 30vh;
@@ -17,9 +17,10 @@
     overflow: hidden;
     background: rgb(44, 44, 44);
     border-radius: 10px;
+    color: black;
     animation: fadeIn .6s;
   }
-  
+
   .header {
     background: rgb(231, 89, 37);
     padding: 1rem;
@@ -29,41 +30,14 @@
   }
 
   .header button{
-    background: none;
-    border: none;
-    cursor: pointer;
+      background: none;
+      border: none;
+      cursor: pointer;
   }
-  
+
   .header h2 {
     margin: 0;
     color: white;
-  }
-  
-  .content {
-    padding: 2rem 1rem;
-    color: white;
-  }
-  
-  .actions {
-    padding: 1rem;
-    display: flex;
-    justify-content: flex-end;
-    gap: .5rem;
-  }
-
-  .actions button{
-      padding: .5rem;
-      width: 7rem;
-      color: white;
-      background: transparent;
-      border: none;
-      border: 1px solid wheat;
-      border-radius: 5px;
-  }
-
-  .actions button:hover{
-      background: rgb(231, 89, 37);
-      cursor: pointer;
   }
 
   @keyframes fadeIn {
@@ -74,4 +48,3 @@
       opacity:1;
     }
   }
-
