+++ src/components/Contacts/NewContactForm.module.css
@@ -1,0 +1,58 @@
+.form input:focus{
+    outline: none;
+}
+
+.form{
+    flex: 1;
+    padding: 1.5rem 0;
+}
+
+.form-control, .submit-btn{
+    padding: .5rem 3rem;
+}
+
+.form-control label{
+    display: block;
+    padding: .3rem 0;
+    color: rgb(97, 96, 96);
+    font-size: 1.1rem;
+}
+
+.form-control input, .form-control select {
+    width: 100%;
+    padding: .5rem;
+    border: none;
+    border-radius: 10px;
+}
+
+.submit-btn button{
+    padding: .5rem;
+    background: none;
+    border: none;
+    font-size: 1.1rem;
+    background: rgb(231, 89, 37);
+    border-radius: 10px;
+    margin-top: .5rem;
+}
+
+.submit-btn button:hover{
+    cursor: pointer;
+    background: rgb(228, 106, 7);
+}
+
+.error-msg{
+    padding: .5rem;
+    color: rgb(167, 19, 19);
+}
+
+.success{
+    padding: .5rem 0;
+    color: rgb(228, 106, 7);
+}
+
+.submit-btn button:disabled,
+.submit-btn button:hover:disabled{
+    background-color: #cccccc;
+    color: #666666;
+    cursor: not-allowed;
+}
